created reserver system

+++ src/services/reserve-service.ts
@@ -12,10 +12,10 @@ export class ReserveService {
   }
 
   static async getReserveById(userId: string, rid: string) {
-    return await reserveRepo.getReserveById(rid);
+    return await reserveRepo.getReserveById(rid, userId);
   }
 
   static async deleteReserve(userId: string, rid: string) {
-    return await reserveRepo.deleteOne({ _id: rid });
+    return await reserveRepo.deleteOne({ _id: rid, userId });
   }
 }
